Expose SRI analysis as an importable library API (#18)
Lift the SRI scanning logic out of the standalone scripts/sri_parser.py
script and into the published package as domain_security_analyzer.sri,
so consumers (e.g. tpcrm-findings-scanner) can reuse it instead of
reimplementing Subresource Integrity parsing.

- Add domain_security_analyzer/sri.py with SRIParser, UnsafeResource, and
  a scan_url() convenience function returning the structured report
  (per-resource integrity/crossorigin values, machine-readable reason
  codes, observed CSP headers, and compensating-control detection).
- Re-export SRIParser/UnsafeResource/scan_url from the package __init__.
- Reduce scripts/sri_parser.py to a thin CLI wrapper that imports the
  core logic, preserving `python scripts/sri_parser.py <url>`.
- Add tests/test_sri.py (network-free) covering each reason code, CSP
  compensating-control logic, and the public API surface.
- Document the importable API in the README SRI section.

Closes the in-repo portion of #18; cross-repo wiring (having
tpcrm-findings-scanner depend on this package and documenting the
detect-vs-dispute boundary in both READMEs) remains tracked there.
